@@ -1,7 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>netcoreapp10.0</TargetFramework>
+    <LangVersion>latest</LangVersion>
+    <Nullable>enable</Nullable>
 
     <License>https://github.com/hbiarge/Acheve.Authentication.Events/blob/master/LICENSE</License>
     <PackageProjectUrl>https://github.com/hbiarge/Acheve.Authentication.Events</PackageProjectUrl>
@@ -11,7 +13,7 @@
       For more information see https://github.com/hbiarge/Acheve.Authentication.Events
     </Description>
     <PackageTags>Authentication</PackageTags>
-    <Version>3.1.3</Version>
+    <Version>4.0.0</Version>
     <RepositoryUrl>https://github.com/hbiarge/Acheve.Authentication.Events</RepositoryUrl>
     <Authors>Hugo Biarge</Authors>
   </PropertyGroup>
@@ -21,13 +23,14 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.Facebook" Version="3.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.Google" Version="3.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.MicrosoftAccount" Version="3.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.OpenIdConnect" Version="3.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.Twitter" Version="3.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.WsFederation" Version="3.1.3" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.Facebook" Version="10.0.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.Google" Version="10.0.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="10.0.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.MicrosoftAccount" Version="10.0.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.OpenIdConnect" Version="10.0.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.Twitter" Version="10.0.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.WsFederation" Version="10.0.2" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.4" />
   </ItemGroup>
 
 </Project>
