@@ -1,125 +1,119 @@
-﻿using System;
+using System;
 using Acheve.Authentication.Events.Local;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using System.Threading.Tasks;
 
-namespace Microsoft.AspNetCore.Authentication
+namespace Microsoft.AspNetCore.Authentication;
+
+public class EventsTypeJwtBearerPostConfigureOptions : IPostConfigureOptions<JwtBearerOptions>
 {
-    public class EventsTypeJwtBearerPostConfigureOptions : IPostConfigureOptions<JwtBearerOptions>
+    public void PostConfigure(string? name, JwtBearerOptions options)
     {
-        public void PostConfigure(string name, JwtBearerOptions options)
+        if (options.EventsType is null)
         {
-            if (options.EventsType is null)
-            {
-                options.EventsType = typeof(LogJwtBearerEvents);
-            }
-            else
-            {
-                options.EventsType = typeof(LogJwtBearerEvents<>).MakeGenericType(options.EventsType);
-            }
+            options.EventsType = typeof(LogJwtBearerEvents);
+        }
+        else
+        {
+            options.EventsType = typeof(LogJwtBearerEvents<>).MakeGenericType(options.EventsType);
         }
     }
+}
 
-    public class LogJwtBearerEvents : JwtBearerEvents
+public class LogJwtBearerEvents(
+    IOptionsMonitor<JwtBearerOptions> options,
+    ILogger<LogJwtBearerEvents> logger) : JwtBearerEvents
+{
+    private readonly IOptionsMonitor<JwtBearerOptions> _options = options;
+    private readonly ILogger<LogJwtBearerEvents> _logger = logger;
+
+    public override async Task Challenge(JwtBearerChallengeContext context)
     {
-        private readonly IOptionsMonitor<JwtBearerOptions> _options;
-        private readonly ILogger<LogJwtBearerEvents> _logger;
+        _logger.Challenge(
+            scheme: context.Scheme.Name,
+            failure: context.AuthenticateFailure,
+            error: context.Error,
+            errorDescription: context.ErrorDescription);
 
-        public LogJwtBearerEvents(IOptionsMonitor<JwtBearerOptions> options, ILogger<LogJwtBearerEvents> logger)
-        {
-            _options = options;
-            _logger = logger;
-        }
+        await SafeCallOriginalEvent(_options.Get(context.Scheme.Name).Events, e => e.Challenge(context));
 
-        public override async Task Challenge(JwtBearerChallengeContext context)
-        {
-            _logger.Challenge(
-                scheme: context.Scheme.Name,
-                failure: context.AuthenticateFailure,
-                error: context.Error,
-                errorDescription: context.ErrorDescription);
-
-            await SafeCallOriginalEvent(_options.Get(context.Scheme.Name).Events, e => e.Challenge(context));
-
-            await base.Challenge(context);
-        }
-
-        public override async Task MessageReceived(MessageReceivedContext context)
-        {
-            _logger.MessageReceived(
-               scheme: context.Scheme.Name);
-
-            await SafeCallOriginalEvent(_options.Get(context.Scheme.Name).Events, e => e.MessageReceived(context));
-
-            await base.MessageReceived(context);
-        }
-
-        public override async Task TokenValidated(TokenValidatedContext context)
-        {
-            _logger.TokenValidated(
-               scheme: context.Scheme.Name,
-               principal: context.Principal,
-               token: context.SecurityToken.Id);
-
-            await SafeCallOriginalEvent(_options.Get(context.Scheme.Name).Events, e => e.TokenValidated(context));
-
-            await base.TokenValidated(context);
-        }
-
-        public override async Task AuthenticationFailed(AuthenticationFailedContext context)
-        {
-            _logger.AuthenticationFailed(
-               scheme: context.Scheme.Name,
-               exception: context.Exception);
-
-            await SafeCallOriginalEvent(_options.Get(context.Scheme.Name).Events, e => e.AuthenticationFailed(context));
-
-            await base.AuthenticationFailed(context);
-        }
-
-        private static async Task SafeCallOriginalEvent(JwtBearerEvents events, Func<JwtBearerEvents, Task> action)
-        {
-            if (events != null)
-            {
-                await action(events);
-            }
-        }
+        await base.Challenge(context);
     }
 
-    public class LogJwtBearerEvents<TOther> : LogJwtBearerEvents where TOther : JwtBearerEvents
+    public override async Task MessageReceived(MessageReceivedContext context)
     {
-        private readonly TOther _originalEvents;
+        _logger.MessageReceived(
+           scheme: context.Scheme.Name);
 
-        public LogJwtBearerEvents(IOptionsMonitor<JwtBearerOptions> options, TOther originalEvents, ILogger<LogJwtBearerEvents> logger)
-        : base(options, logger)
-        {
-            _originalEvents = originalEvents;
-        }
+        await SafeCallOriginalEvent(_options.Get(context.Scheme.Name).Events, e => e.MessageReceived(context));
 
-        public override async Task Challenge(JwtBearerChallengeContext context)
-        {
-            await base.Challenge(context);
-            await _originalEvents.Challenge(context);
-        }
+        await base.MessageReceived(context);
+    }
 
-        public override async Task MessageReceived(MessageReceivedContext context)
-        {
-            await base.MessageReceived(context);
-            await _originalEvents.MessageReceived(context);
-        }
+    public override async Task TokenValidated(TokenValidatedContext context)
+    {
+        _logger.TokenValidated(
+           scheme: context.Scheme.Name,
+           principal: context.Principal,
+           token: context.SecurityToken.Id);
 
-        public override async Task TokenValidated(TokenValidatedContext context)
-        {
-            await base.TokenValidated(context);
-            await _originalEvents.TokenValidated(context);
-        }
+        await SafeCallOriginalEvent(_options.Get(context.Scheme.Name).Events, e => e.TokenValidated(context));
 
-        public override async Task AuthenticationFailed(AuthenticationFailedContext context)
+        await base.TokenValidated(context);
+    }
+
+    public override async Task AuthenticationFailed(AuthenticationFailedContext context)
+    {
+        _logger.AuthenticationFailed(
+           scheme: context.Scheme.Name,
+           exception: context.Exception);
+
+        await SafeCallOriginalEvent(_options.Get(context.Scheme.Name).Events, e => e.AuthenticationFailed(context));
+
+        await base.AuthenticationFailed(context);
+    }
+
+    private static async Task SafeCallOriginalEvent(JwtBearerEvents? events, Func<JwtBearerEvents, Task> action)
+    {
+        if (events is not null)
         {
-            await base.AuthenticationFailed(context);
-            await _originalEvents.AuthenticationFailed(context);
+            await action(events);
         }
+    }
+}
+
+public class LogJwtBearerEvents<TOther>(
+    IOptionsMonitor<JwtBearerOptions> options,
+    TOther originalEvents,
+    ILogger<LogJwtBearerEvents> logger)
+    : LogJwtBearerEvents(options, logger)
+    where TOther : JwtBearerEvents
+{
+    private readonly TOther _originalEvents = originalEvents;
+
+    public override async Task Challenge(JwtBearerChallengeContext context)
+    {
+        await base.Challenge(context);
+        await _originalEvents.Challenge(context);
+    }
+
+    public override async Task MessageReceived(MessageReceivedContext context)
+    {
+        await base.MessageReceived(context);
+        await _originalEvents.MessageReceived(context);
+    }
+
+    public override async Task TokenValidated(TokenValidatedContext context)
+    {
+        await base.TokenValidated(context);
+        await _originalEvents.TokenValidated(context);
+    }
+
+    public override async Task AuthenticationFailed(AuthenticationFailedContext context)
+    {
+        await base.AuthenticationFailed(context);
+        await _originalEvents.AuthenticationFailed(context);
     }
 }
